requirements.txt
# 核心依赖
requests>=2.28.0
python-dotenv>=1.0.0
numpy>=1.24.0

# Gmail 通知（可选）
yagmail>=0.15.0

# 开发依赖
# pytest>=7.0.0
